'Fixes' for NoSectionErrors with allmetrics (SOFTWARE-1388)
First, I load allmetrics.conf later so config sections are defined by
the time allmetrics.conf is loaded. Second, I ignore the contents of
the "[allmetrics args]" section (and give a warning when it is encountered).
Turns out that feature never worked: any entries in that section are passed to
all probes -- most of which error out because they do not recognize that
argument.

I will have to come up with another way to solve the problem of changing
something like the CE type globally.

git-svn-id: https://vdt.cs.wisc.edu/svn/software/rsv/trunk@19017 4e558342-562e-0410-864c-e07659590f8c