Allow RSV probes to specify interval.  SOFTWARE-1798
This adds a new probe option, 'probe-interval', which specifies the
time between runs of the probe.  This allows the admin to have tighter
control over the frequency ("once every five minutes") instead of targeting
a particular time ("15 past the hour").

This change was done for rsv-perfsonar, as it is common for many probes (>200)
to be launched at the same time when using cron-interval.

No probe uses this by default.  However, to enable it, an admin can set:

[org.osg.general.perfsonar-simple]
probe-interval = 600

in the metric file (adjusting the metric name).  If both probe-interval and
cron-interval are present, then probe-interval is preferred.

This makes the probes behave more like the rsv-consumers.



git-svn-id: https://vdt.cs.wisc.edu/svn/software/rsv/trunk@20643 4e558342-562e-0410-864c-e07659590f8c